Turn fix-typechain into hardhat task
