+++ tasks/subtasks.ts
@@ -3,8 +3,8 @@ import {
   TASK_COMPILE_SOLIDITY_GET_ARTIFACT_FROM_COMPILATION_OUTPUT,
 } from "hardhat/builtin-tasks/task-names";
 
-import { subtask, internalTask } from "hardhat/config";
-import { addGasToAbiMethods, setupNativeSolc } from "../utils/tasks";
+import { subtask, task, internalTask } from "hardhat/config";
+import { addGasToAbiMethods, fixTypechain, setupNativeSolc } from "../utils/tasks";
 
 // Injects network block limit (minus 1 million) in the abi so
 // ethers uses it instead of running gas estimation.
@@ -24,5 +24,10 @@ subtask(TASK_COMPILE_SOLIDITY_GET_ARTIFACT_FROM_COMPILATION_OUTPUT)
 
 // Use native solc if available locally at config specified version
 internalTask(TASK_COMPILE_SOLIDITY_COMPILE).setAction(setupNativeSolc);
+
+task("fix-typechain", "Fixes typechain generated types")
+  .setAction(async () => {
+    fixTypechain()
+  })
 
 export {};
