+++ utils/tasks/index.ts
@@ -1,2 +1,3 @@
 export * from "./artifactUtils";
 export * from "./setupNativeSolc";
+export * from "./fixTypechain";
